Move server logic to core
